Fix plugin crash when duplicate paket.template ids in file tree
Each paket.template file in the file tree is spawning a new `PaketPack`
task. This commit handles situation when multiple paket.template files
with the same `id` are location in the file tree.

The plugin will only create one task and skips any other task creation
if a pack task with the same name was already created.