Feature dependency definition in build.gradle (#41)
Description
===========

This is the first part if bringing the basic paket configuration
(paket.dependencies, paket.template, paket.references) into the gradle
project. This change adds custom `paket` dependency handler to the
gradle dependency handler and a nuget repository factory.

*build.gradle*
```groovy
dependencies {
    paket {
        nuget "Test ~> 1"
    }
}

repositories {
   nuget {
       url 'some/url'
   }
}
```

At the moment only `nuget` dependencies are supported.
When dependencies are configured in the `build.gradle` file, a
`paket.dependencies` file will be generated. If this file already
exists, it will be overridden.

Changes
=======

![ADD] paket dependencies handler
![ADD] nuget dependency definition
![ADD] nuget dependency source artifact repository handler
![ADD] new task `paketDependencies`