Filter .asmdef and .asmdef.meta from Paket.Unity cleanup
Description
===========

Unity 2018 adds more control to the script compilation
process with [assembly definition files]. If a user wants to define a
definition file for the dependencies inside the `Paket.Unity3D` directory
she has to recreate it after each install process because the logic in
the `net.wooga.paket-unity` plugin takes ownership of this directory and
might clean it completely. This patch adds a custom pattern to keep all
`.asmdef` and `asmdef.meta` files in the directory structure (at root
and deep nested).

We see this as a middleground solution for projects that want to use
[assembly definition files] or opt out fromi using them.

Changes
=======

![IMPROVE] output cleanup filters out `.asmdef` and `.asmdef.meta`

[assembly definition files]:    https://docs.unity3d.com/Manual/ScriptCompilationAssemblyDefinitionFiles.html