Improve paket.exe caching (#52)
Description
===========

Gradle tends to invalidate it's buildcache whenever it is unclear who
produced specific task outputs. This leads to a lot of redownload
requests for the paket.exe during tests.
This patch adds a custom `upToDateWhen` spec which checks if the
paket.exe is available and if the version is the correct one. I have to
invoke the tool once because it's the fastest cross platform solution.

Changes
=======

* ![IMPROVE] paket.exe caching